requirements.txt
absl-py==1.4.0
aiodns==3.1.1
aiofiles==23.1.0
aiogram==2.25.2
aiogram-utils==0.1.4
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
auto-py-to-exe==2.42.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
bokeh==3.2.2
bottle==0.12.25
bottle-websocket==0.2.9
buildozer==1.5.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cloudpathlib==0.15.1
cloup==0.13.1
cmake==3.27.2
colorama==0.4.6
coloredlogs==15.0.1
colour==0.1.5
comm==0.2.0
comtypes==1.2.0
confection==0.1.3
contourpy==1.1.0
cuda-python==12.2.0
cycler==0.11.0
cymem==2.0.8
Cython==3.0.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dispatcher==1.0
distlib==0.3.7
djangorestframework==3.15.1
dnspython==2.4.2
docutils==0.20.1
EasyProcess==1.1
Eel==0.16.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
entrypoint2==1.1
executing==2.0.1
executor==23.2
fasteners==0.19
fastjsonschema==2.18.1
filelock==3.12.4
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
gevent==23.9.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.40
glcontext==2.4.0
google-api-core==2.12.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
googlemaps==4.10.0
greenlet==3.0.1
grpcio==1.59.0
grpcio-status==1.59.0
gTTS==2.3.2
h5py==3.9.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
inflect==7.0.0
install==1.3.5
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isosurfaces==0.1.0
itsdangerous==2.1.2
jaraco.context==4.3.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.13.1
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
magic-filter==1.0.12
manim==0.17.3
ManimPango==0.4.3
mapbox-earcut==1.0.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.5
mistune==3.0.2
moderngl==5.8.2
moderngl-window==2.4.4
mongoengine==0.24.2
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
multipledispatch==1.0.0
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==2.8.8
nltk==3.8.1
notebook==7.0.5
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pefile==2023.2.7
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pipenv==2023.12.1
platformdirs==3.11.0
playsound==1.3.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
property-manager==3.0
proto-plus==1.22.3
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycairo==1.24.0
pycares==4.4.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.5.1
PyGetWindow==0.0.9
pyglet==2.0.9
Pygments==2.16.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pymongo==4.6.0
PyMsgBox==1.0.9
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
PyRect==0.2.0
pyrr==0.10.3
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-vlc==3.0.18122
pyttsx3==2.90
pytweening==1.0.7
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.2
screeninfo==0.8.1
seaborn==0.12.2
Send2Trash==1.8.2
sh==2.0.6
six==1.16.0
skia-pathops==0.7.4
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
speedtest-cli==2.1.3
sqlparse==0.4.4
srsly==2.4.8
srt==3.5.3
stack-data==0.6.3
streamlit==1.28.1
svgelements==1.9.6
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tinycss2==1.2.1
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.7
validators==0.22.0
verboselogs==1.7
virtualenv==20.24.5
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.9
weasel==0.3.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
whichcraft==0.6.1
widgetsnbextension==4.0.9
wolframalpha==5.0.0
wrapt==1.15.0
xmltodict==0.13.0
xyzservices==2023.7.0
yarl==1.9.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
